sys-utils: add arch(1) back to the official tree
This patch add arch(1) back to util-linux source code tree, but the
command is not installed by defautl.

For more details see "./configure --help".

The arch command is deprecated in favor of "uname -m" (coreutils). The
latest (6.9+) version of coreutils also supports arch(1) as an alias
to "uname -a". Please, if you need arch(1) use the coreutils
implementation.

Signed-off-by: Karel Zak <kzak@redhat.com>