fdisk: check returns in fdisk from partition changes
currently the code in fdisk which changes partition types is a bit fragile ...
it assumes the partition type succeeded instead of checking the user input or
for errors.  ive tweaked the sub functions to return a value indicative of the
functions' success and fdisk now checks/reports based on that.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>