fdisk: calculate +size{K,M,G} in 2^N
fdisk(8) does not calculate partition size (+sizeM or +sizeG)
in MiB or GiB correctly. It uses 10^N instead 2^N.

This patch cleanups +sizeX to:

  +sizeK    -- KiB  (2^10)
  +sizeKB   -- KB   (10^3)
  +sizeM    -- MiB  (2^20)
  +sizeMB   -- MB   (10^6)
  +sizeG    -- GB   (10^9)
  +sizeGB   -- GiB  (2^30)

This patch also fixes the "Last cylinder..." hint message. The "+number"
without any suffix is not a size at all. It's number of cylinders/sectors.
Note, the 10^N suffixes are not proposed to end-uses in the hint message.

Signed-off-by: Karel Zak <kzak@redhat.com>