mount: "can't create lock file" message sometimes means failure, sometimes not
What the patch does is goes from the situation where:

  1) If /etc/mtab doesn't exist and /etc is read-only, you get the
     "can't create lock file" message and the mount fails

  2) If /etc/mtab does exist and /etc is read-only,you get the same
     message but the mount succeeds

Clearly, the failure to update /etc/mtab should either cause the mount
to fail or not ... sometimes causing it to fail, and sometimes not
(each with the same message) is not useful.

This patch sets the same behaviour for create and update mtab. In both
cases it prints error message and the mount succeeds.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Signed-off-by: Karel Zak <kzak@redhat.com>