fdisk: use more readable "GPT" name rather than "EFI GPT"
Although GPT is defined in EFI spec, it is widely used without EFI, as it
reserves space for the MBR and works fine on PC/BIOS systems provided the
bootloader supports it. 

This is in fact the way everyone seems to be following to overcome the 2 TiB
limitation with MSDOS partition table.

Signed-off-by: Robert Millan <rmh@aybabtu.com>