build-sys: add VARSUFFIX to UTIL_CHECK_LIB
The UTIL_CHECK_LIB macro follows the default autoconf behaviour and
generates have_<libname> and HAVE_LIB<LIBNAME> variables.

Some libraries are substitutional (e.g. ncurses and ncursesw). It would be
nice to generate for that libraries the same HAVE_ variables independently
on a library name.

This patch adds optional VARSUFFIX option to UTIL_CHECK_LIB, so the final
variables are have_<varsuffix> and HAVE_LIB<VARSUFFIX>.

For example:

    UTIL_CHECK_LIB(yyy, func)      generates have_yyy and HAVE_LIBYYY
    UTIL_CHECK_LIB(yyy, func, xxx) generates have_xxx and HAVE_LIBXXX

Signed-off-by: Karel Zak <kzak@redhat.com>