blkid: support detection of multiple signatures
The classic way is to return the first successfully detected signature
(filesystem/raid).  Unfortunately, sometimes we need to check for all
possible signatures, because on some volumes (e.g. CD-ROMs) is possible
to store multiple filesystems.

Sometimes we need to check for all filesystems to avoid situation that
we return wrong result (e.g. swap and FAT on the same device).

Signed-off-by: Karel Zak <kzak@redhat.com>