mount: clean up SPEC canonicalization
The SPEC (fsname) field in fstab/mtab could be:

 - devname
 - NAME=value (e.g LABEL, UUID)
 - directory (MS_MOVE, MS_BIND, ..)
 - pseudo-fs keyword (tmpfs, proc, sysfs, ...)

the pseudo-fs keywords shouldn't be canonicalized to absolute path. It
means we have to differ between SPEC and mountpoint (fs_dir).

Unfortunately, the keywords was checked on wrong place. This patch
move this check to the new function canonicalize_spec().

The fsname in mtab entry is canonicalized when the FS type is not
pseudo filesystem.

Signed-off-by: Karel Zak <kzak@redhat.com>