libuuid: fix $libdir in uuid.pc
The -L<dir> gcc option is for directories to be searched for .so/.a
files. It means the directory with development stuff. We have devel
libs in /usr/lib[64].

The side effect is that pkg-config does not return -L with standard
system directories (so make(1) output is more readable and shorter).

Signed-off-by: Karel Zak <kzak@redhat.com>