use right event size for read()
In order to be compatible with future size increases of
the kernel's rfkill structure, userspace should only
read as much as it expects -- the kernel will truncate
the event read if necessary, which is the way we plan
to have compatibility between different versions, should
they ever be necessary.

Thus, the userspace tool needs to use the exact event
size for a read().

If the userspace tool is ever compiled with a newer
kernel then it will need to be adjusted to work with
older kernels, however.