mount: mention mtab for single mount point mount in mount.8
Karel Zak wrote: "This is very old (10+years) feature. The mount(8) is
trying to found mountpoint/device in fstab and then in mtab.

I guess the original purpose for this feature is user-friendly 'mount
/node -o remount'. Unfortunately, the code does not check for
MS_REMOUNT flag and consults mtab in all cases. (The code is
consulting fstab/mtab before mount options parsing.)"

This commit documents this feature.

Signed-off-by: Karel Zak <kzak@redhat.com>