mount: document changed semantics of tmpfs size option in mount.8
Linux kernel commit 818db35992c249dc32c1d86daf7d533fb0952f5d fixed a
problem where invoking mount -t tmpfs with a size argument smaler than the
size of one page caused the limits on the size of a tmpfs to not operate.
The commit also caused the number supplied by the size option to be
rounded up to the  nearest page size instead of down.  This alters the man
page to match the behavior of the kernel in this instance.

Signed-off-by: Kevin Granade <kevin.granade@gmail.com>